statgenie_backend/.ipynb_checkpoints/requirements-checkpoint.txt
flask>=2.2
pandas>=2.1
numpy>=1.24
plotly>=5.20
google-generativeai>=0.6.0
orjson>=3.9.0
werkzeug>=2.2